@@ -1,15 +1,16 @@
 import React from 'react';
-import { HashRouter as HRouter, Route, Redirect } from 'react-router-dom';
+import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import Home from '../Routes/Home';
 import Detail from '../Routes/Detail';
 
 function Router() {
   return (
-    <HRouter>
-      <Route path="/" exact component={Home} />
-      <Route path="/:id" exact component={Detail} />
-      <Redirect path="*" to="/" />
-    </HRouter>
+    <BrowserRouter>
+      <Routes>
+        <Route path="/" element={<Home />} />
+        <Route path="/:id" element={<Detail />} />
+      </Routes>
+    </BrowserRouter>
   );
 }
 
